requirements.txt
xmltodict
python-docx
pytesseract